backend/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
pip==25.0.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
environ==1.0.0
python-dotenv==1.0.1
xlsxwriter==3.1.9
pandas==2.2.3
requests==2.31.0
beautifulsoup4==4.12.3
